+++ apps/commerce-api/src/main/java/com/loopers/application/order/ItemInfos.java
@@ -1,4 +1,4 @@
-package com.loopers.application.order.info;
+package com.loopers.application.order;
 
 import com.loopers.domain.order.orderItem.OrderItemModel;
 import java.math.BigInteger;
