+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderCancelInfo.java
@@ -1,4 +1,4 @@
-package com.loopers.application.order.info;
+package com.loopers.application.order;
 
 import com.loopers.domain.order.OrderModel;
 import com.loopers.domain.order.orderItem.OrderItemModel;
