+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderHistoryHandler.java
@@ -1,10 +1,9 @@
-package com.loopers.application.order.handler;
+package com.loopers.application.order;
 
 import com.loopers.domain.order.OrderModel;
 import com.loopers.domain.order.history.OrderHistoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Component
