+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderItemCommands.java
@@ -1,4 +1,4 @@
-package com.loopers.application.order.command;
+package com.loopers.application.order;
 
 public record OrderItemCommands(
     Long productId,
