+++ apps/commerce-api/src/main/java/com/loopers/application/payment/CardType.java
@@ -1,4 +1,4 @@
-package com.loopers.application.payment.gateway;
+package com.loopers.application.payment;
 
 public enum CardType {
   SAMSUNG,
