+++ apps/commerce-api/src/main/java/com/loopers/application/payment/StockDecreaseCommand.java
@@ -1,4 +1,4 @@
-package com.loopers.application.payment.command;
+package com.loopers.application.payment;
 
 public record StockDecreaseCommand(Long productId, Long quantity) {
 }
