+++ apps/commerce-api/src/main/java/com/loopers/application/payment/TransactionStatusResponse.java
@@ -1,4 +1,4 @@
-package com.loopers.application.payment.gateway;
+package com.loopers.application.payment;
 
 public enum TransactionStatusResponse {
   PENDING,
