+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/PaymentGateway.java
@@ -1,5 +1,7 @@
-package com.loopers.application.payment.gateway;
+package com.loopers.domain.payment;
 
+import com.loopers.application.payment.PaymentRequest;
+import com.loopers.application.payment.PaymentResponse;
 import com.loopers.interfaces.api.ApiResponse;
 
 public interface PaymentGateway {
