+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/payment/PaymentAdapter.java
@@ -1,8 +1,8 @@
 package com.loopers.infrastructure.payment;
 
-import com.loopers.application.payment.gateway.PaymentGateway;
-import com.loopers.application.payment.gateway.PaymentRequest;
-import com.loopers.application.payment.gateway.PaymentResponse;
+import com.loopers.domain.payment.PaymentGateway;
+import com.loopers.application.payment.PaymentRequest;
+import com.loopers.application.payment.PaymentResponse;
 import com.loopers.interfaces.api.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
