+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/payment/PaymentGatewayClient.java
@@ -1,7 +1,7 @@
 package com.loopers.infrastructure.payment;
 
-import com.loopers.application.payment.gateway.PaymentRequest;
-import com.loopers.application.payment.gateway.PaymentResponse;
+import com.loopers.application.payment.PaymentRequest;
+import com.loopers.application.payment.PaymentResponse;
 import com.loopers.interfaces.api.ApiResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
