+++ apps/commerce-api/src/test/java/com/loopers/application/payment/PaymentServiceIntegrationTest.java
@@ -4,12 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.loopers.application.order.OrderFacade;
-import com.loopers.application.order.command.OrderCreateCommand;
-import com.loopers.application.order.command.OrderItemCommands;
-import com.loopers.application.order.info.ItemInfos;
-import com.loopers.application.order.info.OrderCreateInfo;
-import com.loopers.application.payment.command.PaymentCommand;
-import com.loopers.application.payment.gateway.CardType;
+import com.loopers.application.order.OrderCreateCommand;
+import com.loopers.application.order.OrderItemCommands;
+import com.loopers.application.order.ItemInfos;
+import com.loopers.application.order.OrderCreateInfo;
 import com.loopers.application.point.PointFacade;
 import com.loopers.domain.catalog.product.stock.StockModel;
 import com.loopers.domain.catalog.product.stock.StockRepository;
